+++ src/test/java/com/neo4jh3/Neo4jH3Test.java
@@ -122,7 +122,7 @@ public class Neo4jH3Test {
                  }
 
                 result = session.run("RETURN com.neo4jh3.version() AS value");
-                assertEquals("\"5.14.0\"", result.single().get("value").toString());
+                assertEquals("\"5.15.0\"", result.single().get("value").toString());
 
                 result = session.run("RETURN com.neo4jh3.cellToLatLngString('892830926cfffff') AS value");
                 assertEquals("\"37.564248,-122.325306\"", result.single().get("value").toString());